Refactor: 태그 정리 및 컴포넌트 정리
