+++ src/pages/Group/index.js
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { clearHabitDetail } from '../../redux/habitSlice';
 import { useDailyHabits } from '../../hooks/useDailyHabits';
 import getCurrentDate from '../../utils/getCurrentDate';
@@ -10,8 +10,6 @@ import HabitDetailAndVerification from '../../components/habits/HabitDetailAndVe
 function Group() {
   const dispatch = useDispatch();
   const { groupId } = useParams('groupId');
-
-  const habit = useSelector((state) => state.habit.habitDetail);
 
   const currentDate = getCurrentDate();
   const { dailyHabits, loading, error } = useDailyHabits(
@@ -29,7 +27,7 @@ function Group() {
     <section className='min-h-screen flex-1  flex flex-col bg-main-bg text-white bg-vignette'>
       <article className='flex mx-auto mt-28'>
         <HabitList dailyHabits={dailyHabits} />
-        <HabitDetailAndVerification habit={habit} />
+        <HabitDetailAndVerification />
       </article>
     </section>
   );
